Haven't committed in a while... Sorry future Tim and everyone.

+++ heffnormFromCraig/SimpleTypes.h
@@ -1,4 +1,4 @@
 #pragma once
 
-typedef long Index_t; // for nodes, edges and triangles
+typedef int Index_t; // for nodes, edges and triangles
